Revert ESM changes
