+++ src/cli/lint/internalLints/upgrade/patches/16.1.0/index.ts
@@ -1,6 +1,5 @@
 import type { Patches } from '../../index.js';
 
-import { tryAddDdTraceEsmImport } from './addDdTraceEsmImport.js';
 import { tryAddSeekPackageRegistry } from './addSeekPackageRegistry.js';
 import { tryPatchTsconfig } from './patchTsconfig.js';
 import { pruneDevDeps } from './pruneDevDeps.js';
@@ -9,11 +8,6 @@ export const patches: Patches = [
   {
     apply: tryAddSeekPackageRegistry,
     description: 'Add SEEK registry to .npmrc files for SEEK-Jobs repositories',
-  },
-  {
-    apply: tryAddDdTraceEsmImport,
-    description:
-      'Add dd-trace ESM --import to NODE_OPTIONS for Datadog lambdas with handler redirection disabled',
   },
   {
     apply: pruneDevDeps,
